third_party/blink/web_tests/http/tests/websocket/error-event-ready-state-expected.txt

Test that readyState is CLOSED within onerror event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test start (Port 8880)
PASS ws.readyState is 3
PASS successfullyParsed is true

TEST COMPLETE
